Move Zigbe stack related calls inside Zigbee class.
